refactor(selector4nix): rename internal config module to more specific one
