+++ src/infrastructure/config/mod.rs
@@ -1,7 +1,7 @@
-mod parsed;
-mod raw;
+mod general_parsed;
+mod general_raw;
 
-pub use parsed::{
+pub use general_parsed::{
     AppConfiguration, CacheConfiguration, CacheInfoConfiguration, NetworkConfiguration,
     ServerConfiguration, SubstituterConfiguration,
 };
